requirements.txt
requests
google-generativeai
python-dotenv
crewai
crewai-tools
streamlit